NT Fix Linter Rule on Variable Naming
The linter rule is unnecessarily strict on variable and method naming
and do not allow names such as `xAxis` and `yDimension`. There is
nothing on our style guide that would say we should restrict so that
single letter prefixed variable names are not allowed. Relaxing the
rule to reflect that.